[Data] Fix progress bars being displayed as partially completed in Jupyter notebooks (#46289)
<!-- Thank you for your contribution! Please review
https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before
opening a pull request. -->

There is a bug with progress bars generated from Ray Data on Jupyter
notebooks, where the progress bar is left partially complete after the
dataset finishes executing:
![Screenshot at Jun 26
18-09-15](https://github.com/ray-project/ray/assets/5122851/0cb157a8-37fc-4d74-b84a-0831674cfa93)

This PR fixes the bug so that progress bars are marked as fully
completed after execution finishes. We also add a "success" or "failure"
message in the bar after execution terminates:
- On success:
![success](https://github.com/ray-project/ray/assets/5122851/a461366f-e251-44e6-ad9f-4776c6abe22c)
  - On failure: 

![failure](https://github.com/ray-project/ray/assets/5122851/22ed797f-8d76-4101-9b25-032dc821f231)

The progress bar output for the same code outside Jupyter notebook on
master:

![master](https://github.com/ray-project/ray/assets/5122851/2ef01d1b-821e-4647-beb9-8876fa2e95d4)

The same code with this PR:

![after](https://github.com/ray-project/ray/assets/5122851/0706f622-819c-469f-820a-026b0fc58b3a)

- After the progress bar completes, the progress bar is left remaining
in the output:
![Screenshot at Jun 26
19-55-24](https://github.com/ray-project/ray/assets/5122851/303db95b-acdb-46d3-ae40-b7782140ed26)

---------

Signed-off-by: sjl <sjl@anyscale.com>