[doc] auto-generation of class apis (#46243)
After discuss offline with @bveeramani and @angelinalg, this is another
attempt to auto-generate doc apis. Compared to the previous work, this
focuses only on a single class and its methods. It also supports the
common pattern of logical group of methods within a class, which can be
documented in the code itself.

Only the content of the class doc are generated, and users have freedom
to include that content in any structure they want.

The new flow for adding a new class will be:
- Adding the class name into `_autogen.rst`
- Include the class full name, for example, `ray.data.DataIterator` in
any top level rst.

I made an demo in this PR using DataIterator class.

Test:
- CI
- https://anyscale-ray--46243.com.readthedocs.build/en/46243/

Signed-off-by: can <can@anyscale.com>