[core] move cluster_id to GcsClientOptions, and sends RPC to get it if absent. (#46358)
Previously we have `cluster_id` in (Python)GcsClient::Connect as
argument. This arg is never changed throughout lifetime of a GcsClient
and should be placed to GcsClientOptions.

Renamed Python binding `GcsClientOptions.from_gcs_address` to
`GcsClientOptions.create`.

Also, today's C++ GcsClient never requests a cluster_id, so if it's
absent it never populates it in regular GCS requests. Adds a RPC in
Connect() time to get it if absent. To do that, also adds a `timeout`
arg in case GCS is down.

One caveat: if the GcsClient is in GCS itself, the RPC blocks GCS main
thread and it's deadlocked. So in GCS cluster_id *must* be populated
beforehand.