[dashboard] don't prepend "0" or "1" on non streaming logs (#46556)
Dashboard API `api/v0/logs/file` has a peculiar behavior: it prepends an
`1` on each log file. This is normally not a big deal because the logs
were read by humans anyway, but the same API is also used to download
nsight profiles, and in near future structured loggings, where the
prepended `1` can corrupt a machine readable file.

The `1` comes from #25481 for the need to indicate errors in a streaming
HTTP request. Say the dashboard yielded 5 chunks of logs and found
errors in the 6th. It can't set http status to 500 now because the
header has already been transmitted. Hence we prepended a `1` when it's
logs and `0` when it's exception. The char is consumed by the Dashboard
UI Log Viewer (only 1 char) and the CLI (one char for each chunk), but
not by `curl` or Chrome for obvious reasons.

The design itself is kinda OK, as long as the client parses and treats
the leading char. However when you download, you have to give the
pristine data.

In this PR, I added a new query param `format` = `"text" | "leading_1"`.
The former gives just the data; the latter keeps the old behavior,
prepends each chunk with a `1`.

UI Download buttons now use `text`. The `ray logs` CLI and the UI log
viewer still uses `leading_1` to handle exceptions.

Signed-off-by: Ruiyang Wang <rywang014@gmail.com>
Signed-off-by: Ruiyang Wang <56065503+rynewang@users.noreply.github.com>
Co-authored-by: Alan Guo <aguo@aguo.software>