fix docs: unindent blocks to allow copy pasting more easily (#45432)
Inconsistent indents in code samples of the Getting Started section make
it harder to copy paste the code. Since this is the entry point of many
users, I corrected the wrong indents so that they can copy paste the
code samples more easily.

I also removed some indents in logs for consistency.