[core][runtime_env] Allow full path in conda runtime env. (#45550)
#34956 fixed a conda hanging issue by checking existence of the conda
env name. However it only checks by conda env name, not path - if the
user specifies an absolute path, it no longer works. This is an
regression because it should support full paths without a problem.

This PR reads conda CLI for the full paths and names. A special
treatment is done for `base` which is just the conda base prefix; and
others are base conda_prefix/envs/name.

Adds tests for both `base` and full path of it; and also name and full
paths for non base envs.

Signed-off-by: Ruiyang Wang <rywang014@gmail.com>
Co-authored-by: Jiajun Yao <jeromeyjj@gmail.com>