[Data] Add support for objects to Arrow blocks (#45272)
<!-- Thank you for your contribution! Please review
https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before
opening a pull request. -->

<!-- Please add a reviewer to the assignee section when you create a PR.
If you don't have the access to it, we will shortly find a reviewer and
assign them to your PR. -->

## Why are these changes needed?

Currently, Ray does not support blocks/batches with objects and
multi-dimensional arrays in different columns. This causes Ray Data to
throw exceptions when these are provided because:
1. Since there's an arbitrary object in the batch, the Arrow block
format fails with ArrowNotImplemented with dtype 17. This falls back to
`return pd.DataFrame(dict(batch))` in `BlockAccessor.batch_to_block`.
2. However, this particular DataFrame constructor does not support
columns with numpy.ndarray objects, so it throws the exception listed in
the linked issue.

This change enables Python object storage in the Arrow blocks by
defining an Arrow extension type that simply represents the Python
objects as a variable-sized large binary. I suppose the alleged
performance benefits listed in the comments are an extra benefit.

I'm not sure that this is the correct approach or that I've properly
patched all of the places, so some help would be appreciated!

## Related issue number

Resolves #45235
## Checks

- [X] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [X] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [ ] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [X] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(

---------

Signed-off-by: Peter Wang <peter.wang9812@gmail.com>
Signed-off-by: Hao Chen <chenh1024@gmail.com>
Co-authored-by: Peter Wang <peter.wang9812@gmail.com>
Co-authored-by: Hao Chen <chenh1024@gmail.com>