[core][experimental] Add an actionable log when detecting a deadlock caused by using NCCL between DAG nodes on the same actor (#46673)
NCCL is a blocking operation, which means the writer and the reader need to write and read at the same time. However, an actor can only execute one Ray task at a time. Hence, it is impossible for both the reader and writer to operate simultaneously, so the NCCL operation will block forever. If both the reader and the writer are on the same actor, IntraProcessChannel should be used. However, if users specify TorchTensorType, NCCL channel will be created instead.

This PR adds an actionable log when detecting a deadlock caused by using NCCL between DAG nodes on the same actor.