[Data] Remove read task warning if size bytes not set in metadata (#46765)
`Datasource.get_read_tasks()` returns a list of `ReadTask`, where each
`ReadTask` encapsulates a data reading function and it's associated
output metadata. If a serialized `ReadTask` is more than 100KB and
`size_bytes` isn't set in the output metadata (which is often the case),
then Ray Data emits a warning per read task:

```
WARNING: the read task size (288451 bytes) is larger than the reported output size of the task (None bytes). This may be a size reporting bug in the datasource being read from.
```

This warning isn't helpful and usually doesn't indicate an actual issue,
so I'm removing it.

Signed-off-by: Balaji Veeramani <balaji@anyscale.com>