[Data] Truncate progress bar description (#46801)
For ProgressBars used by Ray Data's executor to display operator completion progress, the description of the bar is currently the full operator name. This can become very long and unwieldy with operators with long names, or datasets with many operators (e.g. consecutive MapBatches operators become fused into one giant operator with a really long name).

This PR adds logic to truncate the ProgressBar's description if it exceeds 100 characters. There is also a parameter to disable this truncation, and always show the full progress bar description.

Signed-off-by: Scott Lee <sjl@anyscale.com>