[data] refactor fetching total resources (#46856)
Currently Ray Data's ResourceManager uses the total cluster resources as
available resources. This is problematic when running multiple datasets
or when running Data together with Train. This PR abstracts fetching
total resources as a method in the `Autoscaler` class. So the concrete
Autoscaler implementation can better handle allocating resources.
Also removes the `ignore_extra_tasks` logic in `assert_task_metrics`.
This is not very helpful and can be broken if an Autoscaler
implementation introduces new tasks.


Signed-off-by: Hao Chen <chenh1024@gmail.com>