[ci][tune][py312] Remove rllib dependencies from Tune unit tests (#46905)
Many, many Tune tests currently depend on rllib default trainables
(`"PPO"`, `"__fake"`). RLlib depends on tensorflow, which is not easy to
upgrade for py312. The current policy is to skip tensorflow tests, but
the problem is that many Tune unit tests depend on RLlib + TF even
though they're testing generic, core functionality.

Therefore, the solution should be to decouple rllib from Tune tests, so
that the core Tune tests do not need to be skipped for this py312
effort.

---------

Signed-off-by: Justin Yu <justinvyu@anyscale.com>