Generate Versions JSON for new Version Switcher (#46901)
## Why are these changes needed?

In order to move away from the "Read the Docs" version switcher and to
the one provided by the PyData theme we must first have a stable JSON of
the available versions as described here:
https://github.com/google/google-java-format/releases/download/google-java-format-1.7/google-java-format-1.7-all-deps.jar

The adds helpers to custom directives in order to get the list of git
tags starting from the first one we display in the switcher then sort in
semver order. "Latest" and "master" are added to the top. This is run in
the setup fn to generate the JSON in the static folder. After this is
deployed we can reference it, add the new picker, and hide the previous
one.

## Related issue number

This is part one of a fix for
https://github.com/ray-project/ray/issues/46382 and
https://github.com/ray-project/ray/issues/46189. The follow up will
close these out.


Signed-off-by: cristianjd <cristian.j.derr@gmail.com>