[Data] Improve warning message when read task is large (#46942)
When a read task is large, Ray Data prints a message like this for each
read task:

> WARNING: the read task size (17157648427 bytes) is larger than the
reported output size of the task (None bytes). This may be a size
reporting bug in the datasource being read from.

The message isn't actionable, hard to understand, and spammy (it can be
printed thousands of times). To address these issues, this PR makes Ray
Data emit a warning like this once:

> UserWarning: The serialized size of your read function named 'read_fn'
is 128.0MB. This size relatively large. As a result, Ray might
excessively spill objects during execution. To fix this issue, avoid
accessing `self` or other large objects in 'read_fn'.

---------

Signed-off-by: Balaji Veeramani <balaji@anyscale.com>