[core] GcsClient: No GIL for C++ pb serialization (#46926)
Currently we hold the GIL when we do conversion `C++ pb message` ->
`serialized string` -> `Python pb message`. However only the latter is
needed; the first conversion is a pure C++ call and we should not hold
the GIL.

Adds a `c_vector[c_string]` to hold the strings and do the serialization
in a `nogil` block. Then with gil, we convert it to Python objects.

Not doing per-object locking because GIL is expensive to hold.

Also makes the conversion functions and the future assignment function
`with gil`, so the C++ wrapper code `PyCallback` needs not to do it,
except for error checking.

Signed-off-by: Ruiyang Wang <rywang014@gmail.com>