[core] Fix memleaks in the callback (#47023)
Refcounting in Cython is really pesky, because Cython makes a bunch of +
and - waltz and we need to play along. Per Cython document, the C++
functor PyCallback calls `converter_` and gets a PyObject, should be
responsible to DECREF after it's out of scope, but we didn't. Fixes by
adding the DECREF.

Signed-off-by: Ruiyang Wang <rywang014@gmail.com>