[core] Move UsageStatsClient to gcs_server/. (#47003)
Previously GcsServer keeps a GcsClient to itself. This is awkward and
require careful placement of order of init calls. The dependency is not
needed: the only user is UsageStatsClient which is only used by
GcsServer, so we can move it into gcs_server and let it directly talk to
InternalKV Manager.

Signed-off-by: Ruiyang Wang <rywang014@gmail.com>