[Data] Disable operator fusion for `ReadParquetFragments` and `ReadFiles` physical operators (#881)
The performance of `xgboost_train_batch_inference_benchmark_10G`
regressed because of a bug related to operator fusion and
https://github.com/anyscale/runtime/pull/841:
* `PartitionParquetFragments` outputs 66 blocks containing 3-4 rows each
(260 rows total). Each row contains a fragment, and each block
represents 128 MiB of read data (the actual size of the block is much
smaller)
* `ReadParquetFragments` is fused with XGBoostPredictor, which has a
batch size of 8192
* `ReadParquetFragments `implicitly also uses a batch size of 8192
because of operator fusion
* `ReadParquetFragments` reads the entire dataset (all 260 fragments) in
a single task to satisfy the implicit batch size

This PR disables operator fusion for `ReadParquetFragments` and
`ReadFiles` as an immediate solution. As a follow-up, we should
implement a more holistic fix for issues related to operator fusion with
varying batch sizes.

---------

Signed-off-by: Balaji Veeramani <bveeramani@berkeley.edu>