[core][experimental] Build an operation-based execution schedule for each actor to avoid deadlocks caused by NCCL operations (#46911)
Generate an execution schedule for each actor. The schedule is a list of DAGNodeOperation.
Step 1: Generate a graph based on the following rules:

Divide a DAG node into three GraphNodes: READ, COMPUTE, and WRITE. Each GraphNode has a DAGNodeOperation.
Add edges between READ and COMPUTE, and between COMPUTE and WRITE, which belong to the same task.
Add an edge between COMPUTE with bind_index i and COMPUTE with bind_index i+1 if they belong to the same actor.
Add an edge between WRITE of the writer task and READ of the reader task.
Step 2: Topological sort: If there are multiple GraphNodes with zero in-degree, select one based on the following rules:

(1) If the nodes are not NCCL write nodes, select the one with the smallest bind_index. If there are multiple candidate nodes with the smallest bind_index of the actors that they belong to, any one of them is acceptable. For the implementation details, we maintain a priority queue for each actor, where the peek of the priority queue is the node with the smallest bind_index.
(2) If the node is an NCCL write node, select it only if all of its downstream nodes are also the peeks of their priority queues.
(3) If (1) and (2) cannot be satisfied, it means that all candidate nodes are NCCL write nodes. In this case, select the one that is the peek of the priority queue and its downstream nodes, regardless of whether the downstream nodes are peeks of their priority queues or not.
Then, put the selected nodes into the corresponding actors' schedules.