[Ray backport 2.11] [dashboard] add memray to requirements.txt again (#44401) (#13)
#44315 is reverted since windows tests are failing. The reason is that
memray does not support windows so the pip requirements is failed to
run.

This PR fixed the requirements issue by conditional install memray when
current system platform is not win32

Note: In last PR, we added memray as a required dependency for dashboard
(default), and it's not set required in this PR.

---------

<!-- Thank you for your contribution! Please review
https://github.com/ray-project/ray/blob/master/CONTRIBUTING.rst before
opening a pull request. -->

<!-- Please add a reviewer to the assignee section when you create a PR.
If you don't have the access to it, we will shortly find a reviewer and
assign them to your PR. -->

## Why are these changes needed?

<!-- Please give a short summary of the change and the problem this
solves. -->

## Related issue number

<!-- For example: "Closes #1234" -->

## Checks

- [ ] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [ ] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [ ] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(

Signed-off-by: bhuang <bhuang@anyscale.com>
Co-authored-by: Bryce Huang <bhuang@anyscale.com>
Signed-off-by: Saurabh Vishwas Joshi <sjoshi@sjoshi-7534X5H.dyn.pinadmin.com>