[Data] [Release Test] Use worker node instead of head node for `read_images_comparison_microbenchmark_single_node` (#47228)
## Why are these changes needed?

Recently, there was a change to the OOM killing threshold, being reduced
from 95% to 75% for head nodes. We suspect this to be the root cause of
`read_images_comparison_microbenchmark_single_node` release test
failures, which uses a single head node. To mitigate the issue, this PR
changes the compute config of the test, to instead use a 0-CPU head
node, and a single worker node.

## Related issue number

## Checks

- [x] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [x] I've run `scripts/format.sh` to lint the changes in this PR.
- [ ] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [x] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [ ] Unit tests
- [x] Release tests -
https://buildkite.com/ray-project/release/builds/21386
   - [ ] This PR is not tested :(

---------

Signed-off-by: Scott Lee <sjl@anyscale.com>